backend/requirements.txt
fastapi
uvicorn
pydantic
langchain
langchain-community
langchain-core
groq
pinecone
python-dotenv
python-multipart
transformers[torch]>=4.38
optimum>=1.13.2
onnxruntime>=1.17
onnx